feat(authentication-service): make service for login activity repository functions
GH-1980
