+++ services/authentication-service/src/services/index.ts
@@ -3,10 +3,12 @@
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 import {ActiveUserFilterBuilderService} from './active-user-fliter-builder.service';
+import {LoginActivityHelperService} from './login-activity-helper.service';
 import {LoginHelperService} from './login-helper.service';
 import {OtpService} from './otp.service';
 import {UserViewService} from './user-view.service';
 export * from './active-user-fliter-builder.service';
+export * from './login-activity-helper.service';
 export * from './login-helper.service';
 export * from './otp.service';
 export * from './user-view.service';
@@ -16,4 +18,5 @@ export const services = [
   OtpService,
   ActiveUserFilterBuilderService,
   UserViewService,
+  LoginActivityHelperService,
 ];
